+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @workflow/nuxt
 
+## 5.0.0-beta.19
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/nitro@5.0.0-beta.19
+
+## 5.0.0-beta.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/nitro@5.0.0-beta.18
+
 ## 5.0.0-beta.17
 
 ### Patch Changes